Fix metamask context
