+++ packages/site/src/components/Header.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import styled, { useTheme } from 'styled-components';
-import { MetamaskActions, MetaMaskContext } from '../hooks';
+import { MetamaskActions, MetaMaskContext } from '../contexts/MetamaskContext';
 import { connectSnap, getSnap, getThemePreference } from '../utils';
 import { HeaderButtons } from './Buttons';
 import { SnapLogo } from './SnapLogo';
