+++ packages/site/src/contexts/MetamaskContext.tsx
@@ -1,0 +1,30 @@
+import { createContext, Dispatch } from 'react';
+import { Snap } from '../types';
+
+export type MetamaskState = {
+  isFlask: boolean;
+  installedSnap?: Snap;
+  error?: Error;
+};
+
+export const initialState: MetamaskState = {
+  isFlask: false,
+  error: undefined,
+};
+
+export type MetamaskDispatch = { type: MetamaskActions; payload: any };
+
+export const MetaMaskContext = createContext<
+  [MetamaskState, Dispatch<MetamaskDispatch>]
+>([
+  initialState,
+  () => {
+    /* no op */
+  },
+]);
+
+export enum MetamaskActions {
+  SetInstalled = 'SetInstalled',
+  SetFlaskDetected = 'SetFlaskDetected',
+  SetError = 'SetError',
+}
